Add equivalent policy support for encrypt mode
This adds support for translating labels from a local policy to a remote
one, known as encrypt mode.

Both source and target policy files must be loaded into the Site object for
this to work, and the source policy needs to have the equivalent policy data
present.

Spiffing aims to support the full range of XML SPIF 2.0.